Home Page
I performed an in-depth Power BI analysis on the Tata Group and the diverse car models of the Tata Group, one of the most renowned conglomerates in the world. This project was inspired by the incredible contributions of Mr. Ratan Tata Sir, whose leadership has significantly shaped the Tata Group's legacy.